Move CNodeState and friends out of main.cpp
The purpose of this move is to make it easier to write code outside of main.cpp
